Add atoi function to enable numeric expressions
Even though fields like transaction ID are numeric, they are only
available as strings in the expression engine. Registering an atoi
function enables the use of numeric operators in expressions.

One use case is taking actions with a random chance. For example, say
you want to drop 5% of queries to introduce some chaos. Since the
transaction ID is a random uint16, it can be used to drive this
decision:

```
firewall query {
  # 5% chance to drop (based on transaction id, a random uint16)
  drop atoi([>id]) / 65535.0 < 0.05
  allow true
}
```

Signed-off-by: Christopher Broglie <cbroglie@cloudflare.com>