Add random function
The atoi function was added in #32 to enable policy expressions to make
decisions with a random chance based on the transaction ID. The client
*should* generate unpredictable transaction IDs from a uniform
distribution, but many client libraries reuse the same transaction ID
when retrying a query. Therefore, deterministically dropping queries
based on the transaction ID doesn't work well for simulating packet
loss, as retries will never succeed.

The random function enables policies to take action with random chance,
without depending on the client for the quality of randomness:
```
firewall query {
  # 5% chance to drop
  drop random() < 0.05
  allow true
}
```

Signed-off-by: Christopher Broglie <cbroglie@cloudflare.com>